Configured check mk agent.
Former-commit-id: e42fe5a6aaff6eb0c586f3a87bc2bbcf2157b299 [formerly 23b37fa8a2cd71b5737ae5e7dfa0635a3c4f1571] [formerly abe5673430a1b17c73a5d423bfe56dc7f18bd186 [formerly 2e6efbfa8bca5a2da46c4cc227a2ad0c93fa9d2d] [formerly 2e6efbfa8bca5a2da46c4cc227a2ad0c93fa9d2d [formerly 2e6efbfa8bca5a2da46c4cc227a2ad0c93fa9d2d [formerly 2e6efbfa8bca5a2da46c4cc227a2ad0c93fa9d2d [formerly 2e6efbfa8bca5a2da46c4cc227a2ad0c93fa9d2d [formerly 2e6efbfa8bca5a2da46c4cc227a2ad0c93fa9d2d [formerly ca13567e110bfa518f37ea5761ab9d29719c08d5]]]]]]]
Former-commit-id: abe5673430a1b17c73a5d423bfe56dc7f18bd186 [formerly 2e6efbfa8bca5a2da46c4cc227a2ad0c93fa9d2d]
Former-commit-id: 4db882e025eed4bbeaec1162c9f18d80d03c27b4
Former-commit-id: 520e7498fede8284604617dc2e1aa4835d16e6d6 [formerly d04e703572eecfb1778af963c0a72ec92178cfe0]
Former-commit-id: 01297da4ddfb05ca989254ea93ba1f50c10354e3
Former-commit-id: 0f8a6adb48a58a2d3fcc47f9152a28e26dc69f0f [formerly 0d24783381adb30ecc4c242307ec843c60797705] [formerly c17fea258e8a422d0fb6fd350e2c9e10dcf7eea6 [formerly 32d8bf70368c6b3dfd6a2d32950ff1f2c955e612]]
Former-commit-id: f9cb4dcea2bc76ad5db2c6a3f5b3c03b1db5facc
Former-commit-id: d1779e4cec05e49dbce20e27477e48be748e7cd2

+++ scripts/omd_seup/agent/setup-centos.sh
@@ -1,19 +1,22 @@
 #!/bin/bash
 
-INSTALL_DIR=/opt/omd
+INSTALLATION_DIRECTORY=/opt/omd
+mkdir -p $INSTALLATION_DIRECTORY
+cd $INSTALLATION_DIRECTORY
 
-mkdir -p $INSTALL_DIR
+echo ...........................Installing check mk agent...........................
+if [ -f "/vshare/base-images/jdk/jdk-8u45-linux-x64.rpm" ]; then
+	yum install -y /vshare/base-images/check_mk/agent/check_mk-agent-1.2.4p5-1.noarch.rpm
+else
+	wget http://mathias-kettner.de/download/check_mk-agent-1.2.4p5-1.noarch.rpm
+	yum -y install check_mk-agent-1.2.4p5-1.noarch.rpm
+fi
 
-cd $INSTALL_DIR
-
-wget http://mathias-kettner.de/download/check_mk-agent-1.2.4p5-1.noarch.rpm
-
-yum -y install check_mk-agent-1.2.4p5-1.noarch.rpm
-
+echo ...........................Configuring check mk agent...........................
 chmod 777 /etc/xinetd.d
-
 cp /mnt/gluster/repo/omd-setup/agent/config/check_mk  /etc/xinetd.d/check_mk
 
+echo ...........................Restarting check mk agent...........................
 service xinetd restart
 
 exit
